+++ project/src/drone_pool.h
@@ -16,14 +16,14 @@ namespace csci3081 {
 	float max_speed_;
 	float base_acc_;
 	float weight_cap_;
-	int base_bat_cap_;
+	float base_bat_cap_;
 
 	droneSpecs(std::string model = "Q-36-01",
                float mass = 50.0,
                float speed = 60.0,
                float acc = 4.0,
                float weight = 10.0,
-               int battery = 300)
+               float battery = 300)
                : model_(model),
                  mass_(mass),
                  max_speed_(speed),
